# option to omit 'last' page link from json:api responses