#### Purpose

Allow jsonapi users to opt out of calculating the `last` page link. 

For complex record sets, the `COUNT(*)` query required to determine the total number of pages adds additional overhead which isn't entirely necessary. Clients can still consume all records by continuing to request pages until the 'next' link is `nil`.

In local testing, this changed one of my endpoints as such (with rails 5.2.3)

```
# without patch
Completed 200 OK in 1536ms (Views: 103.6ms | ActiveRecord: 1431.0ms)

# with patch
Completed 200 OK in 994ms (Views: 100.1ms | ActiveRecord: 892.8ms)
```

The difference is all attributable to the performance of the `COUNT(*)` query.

#### Changes

Adds a new configuration `ActiveModelSerializers.config.jsonapi_omit_total_pages`.
Defaults value is `false`, which maintains current behavior.

When this option is `true`:

  1. the `last` page link will always be `nil`
  2. a different method (which does not use `total_pages`) is used to calculate the `next` page

#### Caveats

If a collection contains a number of records which is evenly divisible by the page size, this option will cause the client to see one additional 'ghost' page (due to the way that the `next` page is calculated). That final page will contain 0 records and will have the expected `next: nil` link.

example:

if a record set contains 9 records, and the page size is 3, the 3rd page (records 7-9) will include a `next` link to page 4, which will have 0 records.

#### Related GitHub issues

  - #2149 ensured that all pagination links are always added to each response
  - json-api/json-api#887. that one is still open, but consensus of commenters seems to be that `null` pagination links are acceptable.

#### Additional helpful information

I believe this is compatible with the json:api pagination specification, which says

> Keys MUST either be omitted or have a null value to indicate that a particular link is unavailable.
https://jsonapi.org/format/#fetching-pagination

in other words, servers may provide a `last` link but are not obligated to.