@alexdean 2019-04-12
CI failures seem unrelated to this PR.

## travis failure

on rub 2.5.3 + rails master with the following error

```
/home/travis/.rvm/gems/ruby-2.5.3/gems/bundler-1.17.3/lib/bundler/rubygems_integration.rb:408:in `block (2 levels) in replace_gem': Error loading the 'sqlite3' Active Record adapter. Missing a gem it depends on? can't activate sqlite3 (~> 1.4), already activated sqlite3-1.3.13. Make sure all dependencies are added to Gemfile. (LoadError)
```

https://travis-ci.org/rails-api/active_model_serializers/jobs/519398637

## appveyor failure

```
Bundler could not find compatible versions for gem "bundler":
  In Gemfile:
    bundler (~> 1.6) x86-mingw32
  Current Bundler version:
    bundler (2.0.1)
This Gemfile requires a different version of Bundler.
Perhaps you need to update Bundler by running `gem install bundler`?
Could not find gem 'bundler (~> 1.6)' in any of the relevant sources:
  the local ruby installation
```

https://ci.appveyor.com/project/bf4/active-model-serializers/builds/23811402/job/64lu1fo8a4gh3rqc